Large commit for almost completely removing locking; add sine table
  Split parameters into processing and sync parameters

Processing parameters hold the actual value used for audio generation
and are borrowed mutably in the processing thread. Sync parameters store
their values atomically as floats between 0.0 and 1.0 (the format used
by the host to set them).

The processing thread reads changes in the sync parameters during
processing though a change tracking construct and set any changes to
the processing parameters.

Preset handling has been rewritten to avoid using locks (except for
preset names.) It is not completely finished: old tests need to be
implemented anew and I need to think about if the current concept
makes sense. Maybe the sync parameters and the presets should be
merged into one structure. This would mean users can change presets
as they wish and there won't be any awkward conversion between
presets and sync parameters. It would also mean that users might
be surprised when they close their DAW and open it again and all their
changes are gone - except if changes are persisted to disk.

Unfortunately, this commit means a duplication of structures and
especially of mapping from parameter index to actual parameter.

f32s are used instead of f64s everywhere now. This helps with
parameter value conversions and should aid future SIMD optimizations.

Notes were renamed to voices, and methods were modified.

SineTableLookup sine approximations were added. Unfortunately,
they do not seem to improve performance in their current state,
so they're not actually used for sound generation.

  Other

* Update phase generation (use mul_add, store phase as 0-1 instead
  of 0-TAU. .fract() is used to keep values between 0 and 1, which
  seems to fix an audio issue and might be good for future sin
  function optimizations
* Update parameter utils for less precision loss
* Use possibly more exact TAU constant (maybe PI * 2.0 was precise
  enough though..)
* Use of mul_add in envelope calculation
* Some tests were improved, many were modified

+++ src/common.rs
@@ -2,19 +2,19 @@
 
 /// Number that gets incremented with 1.0 every second
 #[derive(Debug, Copy, Clone, PartialEq)]
-pub struct TimeCounter(pub f64);
+pub struct TimeCounter(pub f32);
 
 #[derive(Debug, Copy, Clone)]
-pub struct Phase(pub f64);
+pub struct Phase(pub f32);
 
 #[derive(Debug, Copy, Clone)]
-pub struct SampleRate(pub f64);
+pub struct SampleRate(pub f32);
 
 #[derive(Debug, Copy, Clone)]
-pub struct TimePerSample(pub f64);
+pub struct TimePerSample(pub f32);
 
 #[derive(Debug, Copy, Clone)]
-pub struct BeatsPerMinute(pub f64);
+pub struct BeatsPerMinute(pub f32);
 
 #[derive(Debug, Copy, Clone, PartialEq)]
 pub enum EnvelopeStage {
